insurance-fraud-detection-graph-backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.121.1
google-auth==2.43.0
google-genai==1.49.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
prometheus-client==0.23.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.49.3
structlog==25.5.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
